+++ src/main/java/fast_food_rest/repository/BasketRepository.java
@@ -4,5 +4,5 @@ package fast_food_rest.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import fast_food_rest.entity.Basket;
 
-public interface BasketRepository extends JpaRepository<Basket, Integer> {
+public interface BasketRepository extends JpaRepository<Basket, Long> {
 }
