+++ src/main/java/fast_food_rest/repository/FoodRepository.java
@@ -6,8 +6,8 @@ import fast_food_rest.entity.Food;
 
 import java.util.List;
 
-public interface FoodRepository extends JpaRepository<Food, Integer> {
+public interface FoodRepository extends JpaRepository<Food, Long> {
     boolean existsByName(String name);
 
-    List<Food> findAllByCategoryId(Integer categoryId);
+    List<Food> findAllByCategoryId(Long categoryId);
 }
